Fix #6: Correct return value when desegmenting
When a proper packet is not found in the begining of a
segment, the whole segment is rejected in the hope the
dissector can synchronize with the packet stream in the
next segment.

Otherwise, some number of complete packets were found or
the last packet in the segment was incomplete and desegmentaion
is needed. In any case, the dissector tells Wireshark that the
segment was ours by returning buf:len().